+++ simpleclient-1.20.1/src/main/java/simpleclient/SimpleClient.java
@@ -47,8 +47,7 @@ public class SimpleClient implements ClientModInitializer {
             if (mc.player != null && mc.level != null) {
                 if (DiscordRPC.INSTANCE.getIngameTimestamp() == null) DiscordRPC.INSTANCE.setIngameTimestamp(Instant.now());
                 if (mc.getSingleplayerServer() == null) {
-                    InetSocketAddress address = (InetSocketAddress) mc.getConnection().getConnection().getRemoteAddress();
-                    return DiscordRPC.activity("Multiplayer", address.getHostString(), DiscordRPC.INSTANCE.getIngameTimestamp());
+                    return DiscordRPC.activity("Multiplayer", mc.getCurrentServer().ip, DiscordRPC.INSTANCE.getIngameTimestamp());
                 } else {
                     String gamemode = mc.player.isCreative() ? "Creative Mode" : mc.player.isSpectator() ? "Spectator Mode" : "Survival Mode";
                     return DiscordRPC.activity("Singleplayer", gamemode, DiscordRPC.INSTANCE.getIngameTimestamp());